+++ common-db/src/test/resources/merged_sql/executor_include_merged.sql
@@ -1,0 +1,37 @@
+drop database IF EXISTS db_test;
+create database db_test;
+use db_test;
+SET FOREIGN_KEY_CHECKS=0;
+-- ----------------------------
+-- Table structure for role
+-- ----------------------------
+DROP TABLE IF EXISTS `role`;
+CREATE TABLE `role` (
+  `ID` int(10) unsigned NOT NULL AUTO_INCREMENT COMMENT '主键ID，由于此表数据不用上传，所以直接使用自动递增',
+  `NAME` varchar(20) DEFAULT NULL,
+  `DESCP` varchar(36) DEFAULT NULL,
+  PRIMARY KEY (`ID`)
+) ENGINE=InnoDB AUTO_INCREMENT=21 DEFAULT CHARSET=utf8 COMMENT='系统用户';
+
+-- include  executor.sql start;
+CREATE TABLE IF NOT EXISTS `RC_CONFIGURATION_DIFINITION`
+(
+   `NAME`                 VARCHAR(100) NOT NULL,
+   `DESCP`                VARCHAR(50),
+   PRIMARY KEY (`NAME`)
+);
+CREATE TABLE IF NOT EXISTS `RC_CONFIGURATION_VALUE`
+(
+   `CONFIG_NAME`          VARCHAR(100),
+   `NAME`                 VARCHAR(50),
+   `DESCP`                VARCHAR(50),
+   VALUE                VARCHAR(300),
+   PRIMARY KEY (`CONFIG_NAME`, `NAME`)
+);
+ALTER TABLE `RC_CONFIGURATION_VALUE` ADD CONSTRAINT FK_REFERENCE_CONFIG_NAME FOREIGN KEY (`CONFIG_NAME`)
+      REFERENCES `RC_CONFIGURATION_DIFINITION` (`NAME`) ON DELETE RESTRICT ON UPDATE RESTRICT;
+-- name : {name} 
+-- time : {time};
+
+-- include  executor.sql end;
+SET FOREIGN_KEY_CHECKS=1;